+++ lib/easypost/services/insurance.rb
@@ -34,4 +34,12 @@ class EasyPost::Services::Insurance < EasyPost::Services::Service
 
     all(params)
   end
+
+  # Refund an Insurance object
+  def refund(id)
+    # TODO: Move beta when endpoint is in GA.
+    response = @client.make_request(:post, "insurances/#{id}/refund", nil, 'beta')
+
+    EasyPost::InternalUtilities::Json.convert_json_to_object(response, MODEL_CLASS)
+  end
 end